# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для автоматического назначения ревьюверов на Pull Request'ы, управления командами и пользователями.",
        "title": "Pull Request Assigner Service",
        "contact": {
            "name": "Матвей Федоров"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/pullRequest/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает пулл реквест и автоматически назначает до 2 ревьюверов из команды автора",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Создать пулл реквест",
                "parameters": [
                    {
                        "description": "PullRequest payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.createPRRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestCreateOutput"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Автор/команда не найдены",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "PR уже существует",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Идемпотентно устанавливает статус пулл реквеста \"MERGED\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Установить статус пулл реквеста \"MERGED\"",
                "parameters": [
                    {
                        "description": "Merge payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.mergePRRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestMergeOutput"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Меняет одного ревьювера на рандомного из его команды",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Переназначить ревьювера",
                "parameters": [
                    {
                        "description": "Reassign payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.reassignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestReassignOutput"
                        }
                    },
                    "400": {
                        "description": "Неверное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR или пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Нарушение правил переназначения",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает или обновляет команду с указанными пользователями",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Создать команду с участниками",
                "parameters": [
                    {
                        "description": "Team payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.addTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamAddOutput"
                        }
                    },
                    "400": {
                        "description": "Команда уже существует или неверное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/get": {
            "get": {
                "description": "Возвращает информацию о команде и ее пользователях",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Получить команду с участниками",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя команды",
                        "name": "team_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamGetOutput"
                        }
                    },
                    "400": {
                        "description": "Неверное имя команды",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Команда не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestCreateOutput": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestCreateOutputPR"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestCreateOutputPR": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestMergeOutput": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestMergeOutputPR"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestMergeOutputPR": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "mergedAt": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestReassignOutput": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestReassignOutputPR"
                },
                "replaced_by": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.PullRequestReassignOutputPR": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamAddOutput": {
            "type": "object",
            "properties": {
                "team": {
                    "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamAddOutputTeam"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamAddOutputTeam": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamOutputMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamGetOutput": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamOutputMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "github_com_MatTwix_Pull-Request-Assigner_internal_service.TeamOutputMember": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "internal_controller_http_v1.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_v1.ErrorBody"
                }
            }
        },
        "internal_controller_http_v1.addTeamRequest": {
            "type": "object",
            "required": [
                "members",
                "team_name"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controller_http_v1.teamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_v1.createPRRequest": {
            "type": "object",
            "required": [
                "author_id",
                "pull_request_id",
                "pull_request_name"
            ],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_v1.mergePRRequest": {
            "type": "object",
            "required": [
                "pull_request_id"
            ],
            "properties": {
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_v1.reassignRequest": {
            "type": "object",
            "properties": {
                "old_user_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_controller_http_v1.teamMember": {
            "type": "object",
            "required": [
                "is_active",
                "user_id",
                "username"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key required for accessing protected endpoints",
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        }
    }
}